+++ src/LibHac/FsSrv/BaseStorageService.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using LibHac.Common;
 using LibHac.Fs;
-using LibHac.FsSrv.Creators;
+using LibHac.FsSrv.FsCreator;
 using LibHac.FsSrv.Impl;
 using LibHac.FsSrv.Sf;
 using LibHac.FsSystem;
