+++ src/LibHac/FsSrv/FsCreator/EmulatedSdCardFileSystemCreator.cs
@@ -3,7 +3,7 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSrv.Creators
+namespace LibHac.FsSrv.FsCreator
 {
     public class EmulatedSdCardFileSystemCreator : ISdCardProxyFileSystemCreator
     {
