+++ src/LibHac/FsSrv/FsCreator/FileSystemCreatorInterfaces.cs
@@ -1,4 +1,4 @@
-﻿namespace LibHac.FsSrv.Creators
+﻿namespace LibHac.FsSrv.FsCreator
 {
     public class FileSystemCreatorInterfaces
     {
