+++ src/LibHac/FsSrv/FsCreator/IMemoryStorageCreator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using LibHac.Fs;
 
-namespace LibHac.FsSrv.Creators
+namespace LibHac.FsSrv.FsCreator
 {
     public interface IMemoryStorageCreator
     {
