+++ src/LibHac/FsSrv/FsCreator/IPartitionFileSystemCreator.cs
@@ -1,7 +1,7 @@
 ﻿using LibHac.Fs;
 using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSrv.Creators
+namespace LibHac.FsSrv.FsCreator
 {
     public interface IPartitionFileSystemCreator
     {
