+++ src/LibHac/FsSrv/FsCreator/ISdCardProxyFileSystemCreator.cs
@@ -1,6 +1,6 @@
 ﻿using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSrv.Creators
+namespace LibHac.FsSrv.FsCreator
 {
     public interface ISdCardProxyFileSystemCreator
     {
