+++ src/LibHac/FsSrv/FsCreator/ISubDirectoryFileSystemCreator.cs
@@ -1,7 +1,7 @@
 ﻿using LibHac.Common;
 using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSrv.Creators
+namespace LibHac.FsSrv.FsCreator
 {
     public interface ISubDirectoryFileSystemCreator
     {
