+++ src/LibHac/FsSrv/FsCreator/ITargetManagerFileSystemCreator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSrv.Creators
+namespace LibHac.FsSrv.FsCreator
 {
     public interface ITargetManagerFileSystemCreator
     {
